+++ apps/backend/src/app.module.ts
@@ -38,6 +38,7 @@ import { DeviantArtModule } from "./deviantart/deviantart.module";
 import { QueueConsumerModule } from "./queue-consumer/queue-consumer.module";
 import { ImageModerationModule } from "./image-moderation/image-moderation.module";
 import { TraitReviewModule } from "./trait-review/trait-review.module";
+import { DeviantartUuidBackfillModule } from "./jobs/deviantart-uuid-backfill/deviantart-uuid-backfill.module";
 import { Request, Response } from "express";
 
 @Module({
@@ -64,6 +65,9 @@ import { Request, Response } from "express";
       graphiql: process.env.GRAPHQL_PLAYGROUND === "true",
       introspection: process.env.GRAPHQL_INTROSPECTION === "true",
       fieldResolverEnhancers: ["guards", "interceptors", "filters"],
+      subscriptions: {
+        "graphql-ws": true,
+      },
       context: ({ req, res }: { req: Request; res: Response }) => ({
         req,
         res,
@@ -230,6 +234,7 @@ import { Request, Response } from "express";
     QueueConsumerModule,
     ImageModerationModule,
     TraitReviewModule,
+    DeviantartUuidBackfillModule,
   ],
 })
 export class AppModule {}